feat: wire the new metrics to the UI
