+++ src/app/pages/Kafka/api/types.ts
@@ -1,0 +1,8 @@
+import { ConfigurationParameters, RangeQuery } from '@rhoas/kafka-management-sdk';
+
+export type NoUndefinedField<T> = {
+  [P in keyof T]-?: NoUndefinedField<NonNullable<T[P]>>;
+};
+export type SafeRangeQuery = NoUndefinedField<RangeQuery>;
+
+export type BasicApiConfigurationParameters = Pick<ConfigurationParameters, 'accessToken' | 'basePath'>;
